feat: 数据字典导出功能增强与Integer溢出修复
主要改进:

1. 数据库列表下拉选择功能:
   - 新增GET /api/export/databases接口
   - 选择数据源后自动查询数据库列表
   - 支持6种数据库类型(MySQL/Oracle/GaussDB/OceanBase/TDSQL/GoldenDB)
   - 前端使用el-select实现可过滤、可创建的下拉选择
   - 自动过滤系统数据库

2. Integer溢出问题修复:
   - 问题: TEXT类型字段长度4294967295超出Integer范围
   - 修复: ColumnDefinition.length/precision/scale改为Long类型
   - 修复: ColumnInfoRow.length/precision改为Long类型
   - 所有适配器使用rs.getLong()替代rs.getInt()
   - 修复SchemaFlattener、SchemaDictionaryParser、DefaultSchemaDiffer类型兼容

3. 前端错误提示优化:
   - 增强导出错误处理逻辑
   - 解析后端JSON错误信息
   - 检查blob类型避免误判
   - 显示详细错误消息和HTTP状态码

4. 部署脚本优化:
   - start.sh使用nohup后台运行
   - 显示PID和日志查看提示
   - build.bat更新端口提示为8999

技术细节:
- Long类型支持最大值9223372036854775807,远超数据库字段长度
- 数据库列表懒加载,避免不必要的查询
- 前端allow-create支持下拉选择和手动输入双重模式

+++ schemasync-backend/src/main/java/com/schemasync/adapter/GoldenDBAdapter.java
@@ -177,17 +177,18 @@ public class GoldenDBAdapter implements DatabaseAdapter {
                     column.setColumnName(rs.getString("COLUMN_NAME"));
                     column.setDataType(rs.getString("DATA_TYPE"));
                     
-                    int length = rs.getInt("CHARACTER_MAXIMUM_LENGTH");
+                    // 使用Long支持超大值
+                    long length = rs.getLong("CHARACTER_MAXIMUM_LENGTH");
                     if (!rs.wasNull()) {
                         column.setLength(length);
                     }
                     
-                    int precision = rs.getInt("NUMERIC_PRECISION");
+                    long precision = rs.getLong("NUMERIC_PRECISION");
                     if (!rs.wasNull()) {
                         column.setPrecision(precision);
                     }
                     
-                    int scale = rs.getInt("NUMERIC_SCALE");
+                    long scale = rs.getLong("NUMERIC_SCALE");
                     if (!rs.wasNull()) {
                         column.setScale(scale);
                     }
