+++ schemasync-backend/src/main/java/com/schemasync/model/dict/ColumnDefinition.java
@@ -20,18 +20,19 @@ public class ColumnDefinition {
     
     /**
      * 长度,字符串类型使用,如VARCHAR(100)的100
+     * 使用Long支持超大值(如TEXT类型的4294967295)
      */
-    private Integer length;
+    private Long length;
     
     /**
      * 精度,数值型字段总位数,如DECIMAL(10,2)的10
      */
-    private Integer precision;
+    private Long precision;
     
     /**
      * 小数位数,如DECIMAL(10,2)的2
      */
-    private Integer scale;
+    private Long scale;
     
     /**
      * 是否允许NULL
@@ -72,12 +73,12 @@ public class ColumnDefinition {
     public void setColumnName(String columnName) { this.columnName = columnName; }
     public String getDataType() { return dataType; }
     public void setDataType(String dataType) { this.dataType = dataType; }
-    public Integer getLength() { return length; }
-    public void setLength(Integer length) { this.length = length; }
-    public Integer getPrecision() { return precision; }
-    public void setPrecision(Integer precision) { this.precision = precision; }
-    public Integer getScale() { return scale; }
-    public void setScale(Integer scale) { this.scale = scale; }
+    public Long getLength() { return length; }
+    public void setLength(Long length) { this.length = length; }
+    public Long getPrecision() { return precision; }
+    public void setPrecision(Long precision) { this.precision = precision; }
+    public Long getScale() { return scale; }
+    public void setScale(Long scale) { this.scale = scale; }
     public Boolean getNullable() { return nullable; }
     public void setNullable(Boolean nullable) { this.nullable = nullable; }
     public Object getDefaultValue() { return defaultValue; }
